requirements.txt
pandas
pyarrow
minio
python-dotenv
google-api-python-client
kaggle